refactor advisory workspace evaluate router
