+++ src/app/router_registry.py
@@ -58,6 +58,9 @@ from app.routers.advisory_workspace_assistant import (
 from app.routers.advisory_workspace_compare import (
     router as advisory_workspace_compare_router,
 )
+from app.routers.advisory_workspace_evaluate import (
+    router as advisory_workspace_evaluate_router,
+)
 from app.routers.advisory_workspace_handoff import (
     router as advisory_workspace_handoff_router,
 )
@@ -408,6 +411,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(bank_demo_proof_packs_router)
     app.include_router(advisory_workspaces_router)
     app.include_router(advisory_workspace_actions_router)
+    app.include_router(advisory_workspace_evaluate_router)
     app.include_router(advisory_workspace_versions_router)
     app.include_router(advisory_workspace_resume_router)
     app.include_router(advisory_workspace_compare_router)
